refactor: frontend - filename and path
