+++ frontend/src/app/global-side-bar/interface.ts
